feat: implement command line interface for GPX export with yargs
